+++ packages/core/src/components/toggle-group/index.ts
@@ -1,2 +1,1 @@
-export * as ToggleGroup from './index.parts';
-export { useToggleGroupContext } from './toggle-group';
+export * from './index.parts';
